chore: change quests history to rating review
